Fixed and simplified FzfAgCustom()
Commands cant detect the mode, so we cant use visual mode. This used
to work work with -range and Get_visual_selection(), but in case a
visual selection wasn't active, the function would get an old visual
selection. For this reason we would never expand <cword> which
resulted in unintended results. This is changed so that the function
searches for the supplied queryparam and if no queryparam is inserted
it expands <cword>.