Changed: Use zlogin instead of zshrc. Add gnome-keyring
Use zlogin instead of zshrc as zlogin only executes once on login. This
means that we have a considerable performance improvement when starting
a new shell.

This commit also replaces keychain with gnome-keyring-daemon for ssh
keys. gnome-keyring-daemon is started in zlogin, but won't ask for a
passphrase until an ssh key is required. keychain would ask for an ssh
key on the start of a new shell for the first time.