Remove zlogin and use zshrc
Tmux should no longer zsh as a login shell so it's safe to use zshrc as
our main shell file.

Also removes adding ssh-add for each new shell in favor of configuring
ssh with:

```
$ cat ~/.ssh/config
AddKeysToAgent yes
```

This only asks for passphrases when a key is required and only starts
the daemon then.